Rename o3de-demo-project to RobotVacuumSample
