+++ CMakeLists.txt
@@ -12,7 +12,7 @@ if(NOT PROJECT_NAME)
     cmake_minimum_required(VERSION 3.20)
     include(cmake/EngineFinder.cmake OPTIONAL)
     find_package(o3de REQUIRED)
-    project(ROS2-Gem-Demo
+    project(RobotVacuumSample
         LANGUAGES C CXX
         VERSION 1.0.0.0
     )
