+++ Gem/Registry/assetprocessor_settings.setreg
@@ -2,13 +2,13 @@
     "Amazon": {
         "AssetProcessor": {
             "Settings": {
-                "ScanFolder ROS2-Gem-Demo/Assets": {
-                    "watch": "@GEMROOT:ROS2-Gem-Demo@/Assets",
+                "ScanFolder RobotVacuumSample/Assets": {
+                    "watch": "@GEMROOT:RobotVacuumSample@/Assets",
                     "recursive": 1,
                     "order": 101
                 },
-                "ScanFolder ROS2-Gem-Demo/Registry": {
-                    "watch": "@GEMROOT:ROS2-Gem-Demo@/Registry",
+                "ScanFolder RobotVacuumSample/Registry": {
+                    "watch": "@GEMROOT:RobotVacuumSample@/Registry",
                     "recursive": 1,
                     "order": 102
                 }
