+++ Gem/robot_vacuum_sample_shared_files.cmake
@@ -7,5 +7,5 @@
 #
 
 set(FILES
-    Source/ROS2-Gem-DemoModule.cpp
+    Source/RobotVacuumSampleModule.cpp
 )
