+++ Platform/Android/android_project.json
@@ -1,7 +1,7 @@
 {
     "Tags": ["Android"],
     "android_settings" : {
-        "package_name" : "org.o3de.ROS2-Gem-Demo",
+        "package_name" : "org.o3de.RobotVacuumSample",
         "version_number" : 1,
         "version_name" : "1.0.0.0",
         "orientation" : "landscape"
