+++ Resources/Platform/iOS/Info.plist
@@ -5,15 +5,15 @@
 	<key>CFBundleDevelopmentRegion</key>
 	<string>en</string>
 	<key>CFBundleDisplayName</key>
-	<string>ROS2-Gem-Demo</string>
+	<string>RobotVacuumSample</string>
 	<key>CFBundleExecutable</key>
-	<string>ROS2-Gem-Demo.GameLauncher</string>
+	<string>RobotVacuumSample.GameLauncher</string>
 	<key>CFBundleIdentifier</key>
-	<string>com.amazon.lumberyard.ROS2-Gem-Demo</string>
+	<string>com.amazon.lumberyard.RobotVacuumSample</string>
 	<key>CFBundleInfoDictionaryVersion</key>
 	<string>6.0</string>
 	<key>CFBundleName</key>
-	<string>ROS2-Gem-Demo</string>
+	<string>RobotVacuumSample</string>
 	<key>CFBundlePackageType</key>
 	<string>APPL</string>
 	<key>CFBundleShortVersionString</key>
